Various changes

+++ MECollection.h
@@ -18,7 +18,7 @@
 @property(nonatomic, retain) MEDatabase *database;
 @property(nonatomic, copy) NSString *fullName;
 @property(nonatomic, copy) NSString *name;
-@property(nonatomic, readonly) unsigned long documentsCount;
+@property(nonatomic, readonly) NSUInteger documentsCount;
 @property(nonatomic, readonly) NSString *description;
 @property(nonatomic, readonly) NSString *namespace;
