+++ MECursor.h
@@ -17,14 +17,14 @@
 
 @property(nonatomic, retain) MECollection *collection;
 @property(nonatomic, copy) NSDictionary *params;
-@property(nonatomic, assign) int skipCount;
-@property(nonatomic, assign) int returnCount;
+@property(nonatomic, assign) NSUInteger skipCount;
+@property(nonatomic, assign) NSUInteger returnCount;
 @property(nonatomic, copy) NSDictionary *order;
 
 -(id)initWithCollection:(MECollection *)aCollection query:(NSDictionary *)params;
 
--(MECursor *)skip:(int)number;
--(MECursor *)limit:(int)number;
+-(MECursor *)skip:(NSUInteger)number;
+-(MECursor *)limit:(NSUInteger)number;
 -(MECursor *)order:(NSDictionary *)ordering;
 
 -(NSArray *)documents;
